[tune] Clean up base ProgressReporter  API (#24010)
`set_start_time()` was not implemented for the progress reporter base class, but it's called in `tune.run()`.

Instead of adding new methods to set runtime arguments, this PR moves to a singular and forward-compatible `setup()` method that defaults to no-op. This way custom reporters can make use of runtime information passed to the reporter, but can choose to ignore it per default.