Don't add elements to _agent_ids
Because `get_agent_ids` returns the actual set, when we add to it, this
modifies it.

This isn't prefered; a check shouldn't be modifying something!

So, we defensively copy in the test.

It's perhaps better for the `.get_agent_ids()` function to return a copy
instead, but I thought this was less controversial.