[Serve] Fix deployment func no args called with python (#24096)
Our current behavior is dropping all args / kwargs for both http and python, if user deployment function doesn't take any input. But in the meantime we didn't throw anything if user tries to invoke the function in python with actual args.

This PR adds this back, and added a bit special handling for http case with in-line comments.