📝 Update structure in development docs (#24377)
This is a small update for the structure of the docs about building Ray from source.

My idea was to isolate steps that are shared and then steps required per platform/system. Also consolidating the instructions to clone with git, install, directory structure, etc.

I'm still figuring out the building steps (installing the dependencies for docs in an M1), but I wanted to start the draft right away.